sort out dssp and idr annotation tests with changes
